Feature #2815 & #3146 : adding transverse file filter - adding a new cache service : cache associated to the current thread - adding new common annotations in order to specify service methods - adding tools to work on annotations - adding a new PasteDetail extension around primary keys (from / to) - adding the possibility to communicate dummy files in order  to take them in account in process management checks - adding annotation associated to an engine that handles simulation action process checks on services - adding dummy handled file Attachment (SimpleDocument) converter - adding dummy handled file Thumbnail converter - centralizing the getting of the MaximumFileSize parameter value - fixing some little problems around the data storage quota (in back-office) - handling document import - migrating ThumbnailService to Spring context - refactoring some code

+++ lib-core/src/main/java/com/silverpeas/form/displayers/WysiwygFCKFieldDisplayer.java
@@ -333,7 +333,7 @@ public class WysiwygFCKFieldDisplayer extends AbstractFieldDisplayer<TextField> 
       out.println(" focusManager.focus();");
       out.
           println(
-          "oEditor.insertHtml('<a href=\"'+url+'\"> <img src=\"'+img+'\" width=\"20\" border=\"0\" alt=\"\"/> '+label+'</a> ');");
+              "oEditor.insertHtml('<a href=\"'+url+'\"> <img src=\"'+img+'\" width=\"20\" border=\"0\" alt=\"\"/> '+label+'</a> ');");
       out.println("}");
 
       // Gallery files exists; javascript functions
@@ -418,9 +418,7 @@ public class WysiwygFCKFieldDisplayer extends AbstractFieldDisplayer<TextField> 
             language);
         try {
           Source source = new Source(new FileInputStream(file));
-          if (source != null) {
-            fieldValueIndex = source.getTextExtractor().toString();
-          }
+          fieldValueIndex = source.getTextExtractor().toString();
         } catch (IOException ioex) {
           SilverTrace.warn("form", "WysiwygFCKFieldDisplayer.index", "form.incorrect_data",
               "File not found " + file + " " + ioex.getMessage(), ioex);
@@ -477,7 +475,7 @@ public class WysiwygFCKFieldDisplayer extends AbstractFieldDisplayer<TextField> 
       String code, String language) {
     try {
       FileRepositoryManager.createAbsolutePath(componentId, dir);
-    } catch (Exception e) {
+    } catch (Exception ignored) {
     }
     String path = getPath(componentId);
     String fileName = getFileName(fieldName, objectId, language);
@@ -514,18 +512,18 @@ public class WysiwygFCKFieldDisplayer extends AbstractFieldDisplayer<TextField> 
       return objectId + "_" + language + "_" + fieldName;
     }
   }
-  
+
   public void move(ForeignPK fromPK, ForeignPK toPK) throws IOException {
     moveOrCopy(fromPK, toPK, false, null);
   }
-  
+
   private void moveOrCopy(ForeignPK fromPK, ForeignPK toPK, boolean copy,
       Map<String, String> oldAndNewFileIds) throws IOException {
     String fromPath = getPath(fromPK.getInstanceId());
     String toPath = getPath(toPK.getInstanceId());
 
     File from = new File(fromPath);
-    if (from != null && from.exists()) {
+    if (from.exists()) {
       List<File> files = (List<File>) FileFolderManager.getAllFile(fromPath);
       for (File file : files) {
         String fileName = file.getName();
@@ -565,13 +563,13 @@ public class WysiwygFCKFieldDisplayer extends AbstractFieldDisplayer<TextField> 
       }
     }
   }
-  
+
   private void changeInstanceId(File file, String from, String to) throws IOException {
     String content = FileUtils.readFileToString(file, Charsets.UTF_8);
     String changed = content.replaceAll("/"+from+"/", "/"+to+"/");
     FileUtils.writeStringToFile(file, changed, Charsets.UTF_8);
   }
-  
+
   private void changeImagePath(File file, String from, String to,
       Map<String, String> oldAndNewFileIds) throws IOException {
     String content = FileUtils.readFileToString(file, Charsets.UTF_8);
@@ -584,19 +582,19 @@ public class WysiwygFCKFieldDisplayer extends AbstractFieldDisplayer<TextField> 
     }
     FileUtils.writeStringToFile(file, content, Charsets.UTF_8);
   }
-  
+
   private String replaceInternalImageId(String content, ForeignPK oldPK, ForeignPK newPK) {
     String from = "/componentId/" + oldPK.getInstanceId() + "/attachmentId/" + oldPK.getId() + "/";
     String to = "/componentId/" + newPK.getInstanceId() + "/attachmentId/" + newPK.getId() + "/";
     return content.replaceAll(from, to);
   }
-  
+
   public void cloneContents(ForeignPK fromPK, ForeignPK toPK, Map<String, String> oldAndNewFileIds)
       throws IOException {
     if (oldAndNewFileIds == null) {
       oldAndNewFileIds = new HashMap<String, String>();
     }
-    
+
     List<SimpleDocument> images =
         AttachmentServiceFactory.getAttachmentService().listDocumentsByForeignKeyAndType(fromPK,
             DocumentType.image, null);
@@ -605,7 +603,7 @@ public class WysiwygFCKFieldDisplayer extends AbstractFieldDisplayer<TextField> 
           AttachmentServiceFactory.getAttachmentService().copyDocument(image, toPK);
       oldAndNewFileIds.put(image.getId(), imageCopyPk.getId());
     }
-    
+
     moveOrCopy(fromPK, toPK, true, oldAndNewFileIds);
   }
 
@@ -614,7 +612,7 @@ public class WysiwygFCKFieldDisplayer extends AbstractFieldDisplayer<TextField> 
     String fromPath = getPath(componentIdFrom);
 
     File from = new File(fromPath);
-    if (from != null && from.exists()) {
+    if (from.exists()) {
       // Verifie si le repertoire de destination existe
       try {
         FileRepositoryManager.createAbsolutePath(componentIdTo, dir);
@@ -660,19 +658,21 @@ public class WysiwygFCKFieldDisplayer extends AbstractFieldDisplayer<TextField> 
       String language) {
     return getPath(componentId) + getFileName(fieldName, objectId, language);
   }
-  
+
   public static void removeContents(ForeignPK pk) {
     String fromPath = getPath(pk.getInstanceId());
-    Collection<File> files =
-        FileUtils.listFiles(new File(fromPath), new PrefixFileFilter(pk.getId() + "_"), null);
-    for (File file : files) {
-      FileUtils.deleteQuietly(file);
+    File directory = new File(fromPath);
+    if (directory.exists()) {
+      Collection<File> files =
+          FileUtils.listFiles(directory, new PrefixFileFilter(pk.getId() + "_"), null);
+      for (File file : files) {
+        FileUtils.deleteQuietly(file);
+      }
     }
   }
-  
+
   private static String getPath(String componentId) {
-    String[] dirs = new String[1];
-    dirs[0] = dir;
+    String[] dirs = {dir};
     return FileRepositoryManager.getAbsolutePath(componentId, dirs);
   }
 }
